Compare Values Across DataFrame Rows

This function compares values across rows of a dataframe using a specified function. It calculates a matrix of comparison results for each pair of columns, then aggregates these results row-wise.

compare.rows(num.rows, num.columns, names.columns, names.rows, data, FUN)

Arguments

num.rows

An integer specifying the number of rows in the dataframe.

num.columns

An integer specifying the number of columns in the dataframe.

names.columns

A character vector of column names for the resulting matrix.

names.rows

A character vector of row names for the resulting matrix.

data

A dataframe containing the data to be compared.

FUN

A function that takes two numeric vectors and returns a numeric vector with the comparison result.

Value

A dataframe with the same number of rows as the input dataframe and columns corresponding to the column names in `names.columns`. Each element represents the aggregated comparison result for each row across the columns of the dataframe.

Examples

df <- data.frame(A = c(1, 2, 3), B = c(3, 2, 1))
compare.rows(num.rows = 3, num.columns = 2, names.columns = c("A", "B"), 
             names.rows = c("Row1", "Row2", "Row3"), data = df, 
             FUN = function(x, y) x - y)
#>       A  B
#> Row1 -2  2
#> Row2  0  0
#> Row3  2 -2