Plot a Heatmap from a Dataframe

This function generates a heatmap from a dataframe where the first column contains row labels and the remaining columns contain numerical values. The heatmap visualizes the data with color gradients, where each cell's color intensity represents the value in the dataframe.

# S3 method for class 'heatmap'
plot(df, file_name = NULL, width = 10, height = 8)

Arguments

df

A dataframe where the first column contains row labels and the remaining columns contain numerical values.

file_name

A character string specifying the path where the heatmap should be saved. If NULL, the heatmap is displayed but not saved.

width

The width of the saved heatmap image in inches.

height

The height of the saved heatmap image in inches.

Value

NULL This function does not return any value but displays or saves a heatmap plot.

Examples

# Example dataframe
df <- data.frame(
  Method_row = c("Dataset_1", "Dataset_2", "Dataset_3", "Dataset_4",
                 "Dataset_5", "Dataset_6", "Dataset_7", "Dataset_8",
                 "Dataset_9", "Dataset_10", "Dataset_11", "Dataset_12",
                 "Dataset_13", "Dataset_14"),
  Model_1 = c(0.00000, 0.77826, 0.63916, 0.25991, 0.02795, 0.16619, 0.00000, 0.03745, 0.21842, 0.00000, 0.64531, 0.65923, 0.95908, 0.75754),
  Model_2 = c(0.00000, 0.00000, 0.64321, 0.07396, 0.04796, 0.02318, 0.65064, 0.06364, 0.29758, 0.00000, 0.42674, 0.00000, 0.46024, 0.00000),
  Model_3 = c(0.98048, 0.74349, 0.07750, 0.91903, 0.00000, 0.19980, 0.94278, 0.40225, 0.62382, 0.26068, 0.63548, 0.92840, 0.81356, 0.84530),
  Model_4 = c(0.35771, 0.36067, 0.77198, 0.87929, 0.58135, 0.00000, 0.98281, 0.00000, 0.28533, 0.55196, 0.83904, 0.59029, 0.42269, 0.23011)
)

# Plot the heatmap
plot_heatmap(df, file_name = "heatmap.pdf", width = 10, height = 8)
#> Error in plot_heatmap(df, file_name = "heatmap.pdf", width = 10, height = 8): could not find function "plot_heatmap"